requirements.txt
django==5.0.6
environs==11.0.0
aiogram==3.10.0
APScheduler==3.10.4
django-jazzmin==3.0.0
asgiref==3.8.1
psycopg2
pytz
python-dateutil